+++ data/statistics_x01.go
@@ -783,7 +783,8 @@ func GetOfficeStatistics(from string, to string) ([]*models.OfficeStatistics, er
 				WHERE leg_id IN (
 					SELECT id FROM leg WHERE match_id IN (
 						SELECT m.id FROM matches m WHERE m.match_type_id = 1
-						AND m.is_finished = 1 AND m.updated_at >= ? AND m.updated_at < ?))
+						AND m.is_finished = 1 AND m.updated_at >= ? AND m.updated_at < ?)
+					AND (leg_type_id = 1 OR leg_type_id IS NULL))
 				GROUP BY leg_id)
 			ORDER BY checkout DESC, leg_id
 		) checkouts